+++ ChangeLog.rst
@@ -11,6 +11,7 @@ Tik Manager Version History
 * **Fix.006** Show Preview folder issue fixed for sub-projects
 * **New.007** New settings item for setting frame ranges while referencing scene. Default "Ask" behaviour can now be changed to "Never" and "Always". "Always" makes sense for lighting artists who references animation scenes and "Never" makes sense for animators who only reference rigs and assets.
 * **Fix.008** Ingest scene as right click item hidden in reference mode
+* **Fix.009** Check for updates (Help => Check for updates) command is up and working
 
 3.0
 ---